refactor: 빈티지 적용 (#462)
* refactor: builder 패턴 private로 수정

* refactor: wine dto 패키지 수정

* feat: 와인 빈티지 Entity 및 와인 빈티지 평균 업데이트 기능 구현

* feat: Wine, TastinNote 엔티티에 WineVintage 연관관계 반영

* feat: 와인 빈티지 이벤트 dto 구현

* refactor: 와인 패키징 수정

* test: 패키징 수정에 따른 import 변경

* feat: WineVintage Statistics 및 top3 nose 조회 메서드 구현

* refactor: TastingNote에 빈티지 적용
- TastingNote request/response DTO 수정
- TastingNote 연관관계 vintage 적용
- 연관관계 수정에 따른 Service 수정

* test: 패키징 수정에 따른 import 변경

* feat: 빈티지 에러 코드 추가

* refactor: 테이스팅 노트 request에서 vintageYear NotNull 제거 (#460)

* test: 테이스팅 노트 테스트에 빈티지 적용 (#460)

* test: 와인 테스트에 빈티지 적용 (#460)

* feat: 테이스팅 노트 저장 시 와인, 와인빈티지 평균값 조정 (#460)

* refactor: 와인 저장 시 빈티지도 저장되게 수정

* refactor: 와인빈티지 노트 업데이트 메서드 수정

* refactor: 보유 와인에 빈티지 적용

* refactor: 와인 위시리스트 빈티지 적용 (#460)

* refactor: 테이스팅 노트 저장 시 빈티지 적용 (#460)

* refactor: wineName, vintageYear 분리 (#460)

* refactor: tastingNote와 wine을 join하여 가져오게 수정 (#460)

* refactor: 와인 상세정보, 와인 리뷰 조회 API에 빈티지 적용

* refactor: prinln 제거

* refactor: 보유 와인 dto에 빈티지 추가

* refactor: dto 테스트 코드 삭제

* refactor: WineWishlistTest 빈티지 적용 (#460)

* refactor; WineWishlistContrllerTest 빈티지 적용 (#460)

* test: 보유와인 controller 테스트 빈티지 적용 (#460)

* test: MyWineRepositoryTest 빈티지 적용 (#460)

* test: MyWineServiceImplTest 빈티지 적용

* test: TastingNoteControllerTest 빈티지 적용 (#460)

* test: WineWishlistServiceImplTest 빈티지 적용 (#460)

* test: WineControllerTest 빈티지 적용

* test: WineServiceImplTest 빈티지 적용

* test: WineRepositoryImplTest 빈티지 적용 (#460)

* test: TastingNoteServiceImplTest 빈티지 적용 (#460)

* test: TastingNoteRepositoryTest 빈티지 적용 (#460)

* refactor: Controller 설명 수정 및 변경된 DTO 적용

+++ src/main/java/com/drinkeg/drinkeg/domain/wineWishlist/service/WineWishlistServiceImpl.java
@@ -1,13 +1,13 @@
 package com.drinkeg.drinkeg.domain.wineWishlist.service;
 
 import com.drinkeg.drinkeg.domain.wine.dto.response.WinePreviewResponse;
+import com.drinkeg.drinkeg.domain.wine.wineVintage.domain.WineVintage;
+import com.drinkeg.drinkeg.domain.wine.wineVintage.repository.WineVintageRepository;
 import com.drinkeg.drinkeg.global.apipayLoad.code.status.ErrorStatus;
 import com.drinkeg.drinkeg.domain.member.domain.Member;
-import com.drinkeg.drinkeg.domain.wine.domain.Wine;
 import com.drinkeg.drinkeg.domain.wineWishlist.domain.WineWishlist;
 import com.drinkeg.drinkeg.global.exception.GeneralException;
 import com.drinkeg.drinkeg.domain.member.repostitory.MemberRepository;
-import com.drinkeg.drinkeg.domain.wine.repository.WineRepository;
 import com.drinkeg.drinkeg.domain.wineWishlist.repository.WineWishlistRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -19,19 +19,20 @@ import java.util.List;
 public class WineWishlistServiceImpl implements WineWishlistService{
     private final WineWishlistRepository wineWishlistRepository;
     private final MemberRepository memberRepository;
-    private final WineRepository wineRepository;
+    private final WineVintageRepository wineVintageRepository;
 
     @Override
-    public Long createWineWishlist(Long wineId, String username) {
+    public Long createWineWishlist(Long wineId, Integer vintageYear, String username) {
         Member member = memberRepository.findByUsername(username)
                 .orElseThrow(() -> new GeneralException(ErrorStatus.MEMBER_NOT_FOUND));
-        Wine wine = wineRepository.findById(wineId)
-                .orElseThrow(() -> new GeneralException(ErrorStatus.WINE_NOT_FOUND));
+        if (vintageYear == null) vintageYear = 0;
+        WineVintage wineVintage = wineVintageRepository.findByWineIdAndVintageYear(wineId, vintageYear);
+        if(wineVintage == null) throw new GeneralException(ErrorStatus.WINE_NOT_FOUND);
 
-        if (wineWishlistRepository.existsByMemberAndWine(member, wine))
+        if (wineWishlistRepository.existsByMemberAndWineVintage(member, wineVintage))
             throw new GeneralException(ErrorStatus.WINE_WISHLIST_ALREADY_EXISTS);
 
-        WineWishlist wineWishlist = wineWishlistRepository.save(WineWishlist.create(member, wine));
+        WineWishlist wineWishlist = wineWishlistRepository.save(WineWishlist.create(member, wineVintage));
         return wineWishlist.getId();
     }
 
@@ -43,17 +44,18 @@ public class WineWishlistServiceImpl implements WineWishlistService{
         List<WineWishlist> wishlistWineList = wineWishlistRepository.findByMemberOrderByCreatedAtDesc(member);
 
         return wishlistWineList.stream().map(wineWishlist
-                -> WinePreviewResponse.of(wineWishlist.getWine())).toList();
+                -> WinePreviewResponse.of(wineWishlist.getWineVintage())).toList();
     }
 
     @Override
-    public void deleteWineWishlist(Long wineId, String username) {
+    public void deleteWineWishlist(Long wineId, Integer vintageYear,String username) {
         Member member = memberRepository.findByUsername(username)
                 .orElseThrow(() -> new GeneralException(ErrorStatus.MEMBER_NOT_FOUND));
-        Wine wine = wineRepository.findById(wineId)
-                .orElseThrow(() -> new GeneralException(ErrorStatus.WINE_NOT_FOUND));
+        if (vintageYear == null) vintageYear = 0;
+        WineVintage wineVintage = wineVintageRepository.findByWineIdAndVintageYear(wineId, vintageYear);
+        if(wineVintage == null) throw new GeneralException(ErrorStatus.WINE_NOT_FOUND);
 
-        WineWishlist wineWishlist = wineWishlistRepository.findWineWishlistByMemberAndWine(member, wine).orElseThrow(
+        WineWishlist wineWishlist = wineWishlistRepository.findWineWishlistByMemberAndWineVintage(member, wineVintage).orElseThrow(
                 () -> new GeneralException(ErrorStatus.WINE_WISHLIST_NOT_FOUND));
 
         wineWishlistRepository.deleteById(wineWishlist.getId());
